requirements.txt
discord.py==0.16.12
praw==6.0.0
tweepy==3.7.0
bs4==0.0.1
dashtable==1.4.5
feedparser==5.2.1
Flask==1.0.2
